Add required jack binaries to repo
This is purely so that the CI system can access them in order to bundle them with the pyinstaller-generated .exe file.

Using v1.9.11 of jack, using the binaries from the official installer.

It's a horrid solution. Really we want to be able to install jack properly in the CI system, but for some reason the jack chocolatey package fails to install and I can't find copies of the jack binaries anywhere (the official download comes as a self-extracting exe file).